refactor: migrate to 3-layer Clean Architecture
Simplified architecture from 6 layers to 3 layers:
- UI Layer (Streamlit, Presenter, ViewModel)
- Application Layer (Use Case, DTO)
- Domain Layer (Models, Services, Acoustics)

Changes:
- Remove Adapter layer (direct Use Case usage)
- Consolidate domain modules into unified structure
- Implement Use Case pattern for application logic
- Move constants to appropriate layers
