--- src/acoustics/__init__.py
@@ -1,13 +1,0 @@
-"""Acoustics layer for psychoacoustic models."""
-
-from .roughness import (
-    calculate_dissonance_curve,
-    calculate_roughness_pair,
-    critical_bandwidth,
-)
-
-__all__ = [
-    "calculate_dissonance_curve",
-    "calculate_roughness_pair",
-    "critical_bandwidth",
-]
