+++ src/application/__init__.py
@@ -1,0 +1,5 @@
+"""アプリケーション層。
+
+このパッケージは、ビジネスユースケースを調整するApplication層を提供します。
+UI層とDomain層の橋渡しを担当します。
+"""
