+++ src/domain/__init__.py
@@ -1,11 +1,12 @@
 """Domain layer for xenharmonic music theory models."""
 
-from .harmonics import Harmonic, HarmonicSeries, SawtoothTimbre, TimbreModel
-from .tuning import TuningSystem
+from .models import Harmonic, HarmonicSeries, RoughnessPairResult, SawtoothTimbre, TuningSystem
+from .protocols import TimbreModel
 
 __all__ = [
     "Harmonic",
     "HarmonicSeries",
+    "RoughnessPairResult",
     "SawtoothTimbre",
     "TimbreModel",
     "TuningSystem",
