--- src/services/__init__.py
@@ -1,8 +1,0 @@
-"""Service layer for consonance calculations."""
-
-from .consonance_service import ConsonanceCalculator, ConsonanceResult
-
-__all__ = [
-    "ConsonanceCalculator",
-    "ConsonanceResult",
-]
